requirements.txt
flask
pyyaml
requests
openai
